+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - feat: admin users can remove any account but their own
   - feat: daily log rotation
   - feat: filter layers by {fragment|data}_type
+  - feat: get number of annotations in a layer
+  - feat: get number of media in a corpys
   - BREAKING: getting a queue no longer returns its elements (use new picking routes instead)
   - BREAKING: only admin users can create queues
   - BREAKING: queue push returns "success" message instead of queues